refactor(app-shell): top task timer 結線を useTopTaskTimer に集約 (refs #152)
AppShell.tsx 責務分離 (issue #152) の Chunk 4 (最終)。挙動は変えない。

- `useTopTaskTimer`: useTaskTimer の上に被せて TopTimerBinding に整形 + pause
  modal の bridge state (pauseModalOpen / handlePauseSelect) を内包する hook。
- AppShell では `pendingTasks[0] ?? null` を渡して `topTimer` /
  `pauseModalOpen` / `handlePauseSelect` を分割代入で受け取る。
- 不要になった import (useTaskTimer / PauseReason) を AppShell から除去。

AppShell.tsx は 378 → 351 行。issue #152 起票時の 889 行から最終的に 351 行
(約 -60%)。
